refactor: 두 개의 파일을 한 파일에서 관리하기위해 통합 && 사용경로 수정
